requirements.txt
requests~=2.31.0
folium~=0.15.0
Pillow~=10.2.0
